// src/types/generated/ServiceInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ServiceInfo = { name: string, description: string, loadState: string, activeState: string, subState: string, 
/**
 * enabled / disabled / static / masked / unknown
 */
enabledState: string, };